Amend API to fix rate-limiting issues
Previously, updates were pushed to the Hue hub as soon as they happened. This could cause issues with updates being pushed more frequently than the hub could handle.

This amend changes this behaviour so that a thread is responsible for managing updates to the hub. The thread pushes updates, and then waits an amount of time proportional to the number of updates that were made, in a loop. This ensures that the hub has enough time to process the updates.

Another benefit of this behaviour is that consumers of the API do not need to call apply() to push their changes, this is all handled internally within the API.