Fix issue whereby sometimes lights would show as off in group image
This bug was caused because GroupView was reacting to an observable property change, and reading viewModel?.individualLightViewModels to get the lights.

However, viewModel?.individualLightViewModels was being updated on the same observable property change. This was happening later than GroupView reacting to the change, so the viewModel?.individualLightViewModels still had the old values when it was looked at.